change type of the return value on long

+++ studyc/src/ex.h
@@ -1,7 +1,7 @@
 #ifndef EX_H
 # define EX_H
 int	ft_ex01(void);
-int	ft_find_max(void);
+long	ft_find_max(void);
 long	ft_atoi(char *str);
 int	ft_length(char *str);
 int	ft_ex03(int argc, char *argv[]);
